worker/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
opencv-python-headless>=4.8.0
pydantic>=2.0.0
python-escpos>=3.0
redis>=5.0.0
pygame>=2.5.0
requests>=2.31.0
